+++ src/icalendar/parser_tools.py
@@ -1,10 +1,11 @@
-from typing import Any
+from typing import Any, Union
 
 SEQUENCE_TYPES = (list, tuple)
 DEFAULT_ENCODING = 'utf-8'
+ICAL_TYPE = Union[str, bytes]
 
 
-def from_unicode(value: Any, encoding='utf-8') -> bytes:
+def from_unicode(value: ICAL_TYPE, encoding='utf-8') -> bytes:
     """
     Converts a value to bytes, even if it already is bytes
     :param value: The value to convert
@@ -21,7 +22,7 @@ def from_unicode(value: Any, encoding='utf-8') -> bytes:
     return value
 
 
-def to_unicode(value, encoding='utf-8'):
+def to_unicode(value: ICAL_TYPE, encoding='utf-8') -> str:
     """Converts a value to unicode, even if it is already a unicode string.
     """
     if isinstance(value, str):
@@ -34,7 +35,7 @@ def to_unicode(value, encoding='utf-8'):
     return value
 
 
-def data_encode(data, encoding=DEFAULT_ENCODING):
+def data_encode(data: Union[ICAL_TYPE, dict, list], encoding=DEFAULT_ENCODING) -> bytes:
     """Encode all datastructures to the given encoding.
     Currently unicode strings, dicts and lists are supported.
     """