+++ CHANGES.rst
@@ -65,6 +65,9 @@ New features:
 
 - Test compatibility with Python 3.12
 - Add function ``icalendar.use_pytz()``.
+- Allows selecting components with ``walk(select=func)`` where ``func`` takes a
+  component and returns ``True`` or ``False``.
+- Add compatibility to :rfc:`7529`, adding ``vMonth`` and ``vSkip``
 - Add ``sphinx-autobuild`` for ``livehtml`` Makefile target.
 - Add pull request preview on Read the Docs, building only on changes to documentation-related files.
 - Add link to pull request preview builds in the pull request description only when there are changes to documentation-related files.